Refactor auth into openstack package
